+++ src/OpenTelemetry/Logs/OpenTelemetryLoggerProvider.cs
@@ -19,6 +19,7 @@
 using System;
 using System.Collections;
 using System.Diagnostics;
+using System.Text;
 using System.Threading;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
@@ -70,14 +71,10 @@ namespace OpenTelemetry.Logs
         {
         }
 
-        // Note: This is only for tests. Options will be missing ServiceCollection & ServiceProvider features will be unavailable.
-        internal OpenTelemetryLoggerProvider(OpenTelemetryLoggerOptions options)
-            : this(options, serviceProvider: null, ownsServiceProvider: false)
-        {
-        }
-
         internal OpenTelemetryLoggerProvider(OpenTelemetryLoggerOptions options, IServiceProvider? serviceProvider, bool ownsServiceProvider)
         {
+            OpenTelemetrySdkEventSource.Log.OpenTelemetryLoggerProviderEvent("Building OpenTelemetryLoggerProvider.");
+
             Guard.ThrowIfNull(options);
 
             this.IncludeScopes = options.IncludeScopes;
@@ -91,33 +88,11 @@ namespace OpenTelemetry.Logs
                 Debug.Assert(this.ownedServiceProvider != null, "ownedServiceProvider was null");
             }
 
-            // Step 1: Add any processors added to options.
-
-            foreach (var processor in options.Processors)
-            {
-                this.AddProcessor(processor);
-            }
-
-            this.ResourceBuilder = options.ResourceBuilder ?? ResourceBuilder.CreateDefault();
-
-            if (serviceProvider != null)
-            {
-                // Step 2: Look for any Action<IServiceProvider,
-                // OpenTelemetryLoggerProvider> configuration actions registered and
-                // execute them.
-
-                var registeredConfigurations = serviceProvider.GetServices<Action<IServiceProvider, OpenTelemetryLoggerProvider>>();
-                foreach (var registeredConfiguration in registeredConfigurations)
-                {
-                    registeredConfiguration?.Invoke(serviceProvider, this);
-                }
-            }
+            this.ResourceBuilder ??= ResourceBuilder.CreateDefault();
 
             var configurationActions = options.ConfigurationActions;
             if (configurationActions?.Count > 0)
             {
-                // Step 3: Execute any configuration actions.
-
                 if (serviceProvider == null)
                 {
                     throw new InvalidOperationException("Configuration actions were registered on options but no service provider was supplied.");
@@ -130,22 +105,14 @@ namespace OpenTelemetry.Logs
                     configurationActions[i](serviceProvider, this);
                 }
 
+                OpenTelemetrySdkEventSource.Log.OpenTelemetryLoggerProviderEvent($"Number of actions configured = {configurationActions.Count}.");
                 options.ConfigurationActions = null;
-            }
-
-            if (serviceProvider != null)
-            {
-                // Step 4: Look for any processors registered directly with the service provider.
-
-                var registeredProcessors = serviceProvider.GetServices<BaseProcessor<LogRecord>>();
-                foreach (BaseProcessor<LogRecord> processor in registeredProcessors)
-                {
-                    this.AddProcessor(processor);
-                }
             }
 
             this.Resource = this.ResourceBuilder.Build();
             this.ResourceBuilder = null;
+
+            OpenTelemetrySdkEventSource.Log.OpenTelemetryLoggerProviderEvent("OpenTelemetryLoggerProvider built successfully.");
         }
 
         internal IExternalScopeProvider? ScopeProvider { get; private set; }
@@ -212,6 +179,7 @@ namespace OpenTelemetry.Logs
         /// </remarks>
         public bool ForceFlush(int timeoutMilliseconds = Timeout.Infinite)
         {
+            OpenTelemetrySdkEventSource.Log.OpenTelemetryLoggerProviderForceFlushInvoked(timeoutMilliseconds);
             return this.Processor?.ForceFlush(timeoutMilliseconds) ?? true;
         }
 
@@ -227,25 +195,43 @@ namespace OpenTelemetry.Logs
         /// <returns>The supplied <see cref="OpenTelemetryLoggerOptions"/> for chaining.</returns>
         public OpenTelemetryLoggerProvider AddProcessor(BaseProcessor<LogRecord> processor)
         {
+            OpenTelemetrySdkEventSource.Log.OpenTelemetryLoggerProviderEvent("Started adding processor.");
+
             Guard.ThrowIfNull(processor);
 
             processor.SetParentProvider(this);
 
+            StringBuilder processorAdded = new StringBuilder();
+
             if (this.threadStaticPool != null && this.ContainsBatchProcessor(processor))
             {
+                OpenTelemetrySdkEventSource.Log.OpenTelemetryLoggerProviderEvent("Using shared thread pool.");
+
                 this.threadStaticPool = null;
             }
 
             if (this.Processor == null)
             {
+                processorAdded.Append("Setting processor to ");
+                processorAdded.Append(processor);
+
                 this.Processor = processor;
             }
             else if (this.Processor is CompositeProcessor<LogRecord> compositeProcessor)
             {
+                processorAdded.Append("Adding processor ");
+                processorAdded.Append(processor);
+                processorAdded.Append(" to composite processor");
+
                 compositeProcessor.AddProcessor(processor);
             }
             else
             {
+                processorAdded.Append("Creating new composite processor with processor ");
+                processorAdded.Append(this.Processor);
+                processorAdded.Append(" and adding new processor ");
+                processorAdded.Append(processor);
+
                 var newCompositeProcessor = new CompositeProcessor<LogRecord>(new[]
                 {
                     this.Processor,
@@ -254,6 +240,8 @@ namespace OpenTelemetry.Logs
                 newCompositeProcessor.AddProcessor(processor);
                 this.Processor = newCompositeProcessor;
             }
+
+            OpenTelemetrySdkEventSource.Log.OpenTelemetryLoggerProviderEvent($"Completed adding processor = \"{processorAdded}\".");
 
             return this;
         }