+++ src/OpenTelemetry/Metrics/AggregationType.cs
@@ -76,16 +76,3 @@ namespace OpenTelemetry.Metrics
         HistogramSumCountMinMax = 9,
     }
 }
-
-#pragma warning disable SA1649 // File name should match first type name
-internal static class AggregationTypeMethods
-#pragma warning restore SA1649 // File name should match first type name
-{
-    public static bool IsHistogram(this AggregationType aggType)
-    {
-        return aggType == AggregationType.Histogram
-            || aggType == AggregationType.HistogramMinMax
-            || aggType == AggregationType.HistogramSumCount
-            || aggType == AggregationType.HistogramSumCountMinMax;
-    }
-}