+++ src/OpenTelemetry/Metrics/Builder/MeterProviderBuilderSdk.cs
@@ -14,14 +14,31 @@
 // limitations under the License.
 // </copyright>
 
+#nullable enable
+
 using System.Text.RegularExpressions;
+using Microsoft.Extensions.DependencyInjection;
 
 namespace OpenTelemetry.Metrics
 {
-    internal class MeterProviderBuilderSdk : MeterProviderBuilderBase
+    internal sealed class MeterProviderBuilderSdk : MeterProviderBuilderBase
     {
         private static readonly Regex InstrumentNameRegex = new(
             @"^[a-zA-Z][-.\w]{0,62}$", RegexOptions.IgnoreCase | RegexOptions.Compiled);
+
+        public MeterProviderBuilderSdk()
+        {
+        }
+
+        public MeterProviderBuilderSdk(IServiceCollection services)
+            : base(services)
+        {
+        }
+
+        public MeterProviderBuilderSdk(MeterProviderBuilderState state)
+            : base(state)
+        {
+        }
 
         /// <summary>
         /// Returns whether the given instrument name is valid according to the specification.
@@ -29,7 +46,7 @@ namespace OpenTelemetry.Metrics
         /// <remarks>See specification: <see href="https://github.com/open-telemetry/opentelemetry-specification/blob/main/specification/metrics/api.md#instrument"/>.</remarks>
         /// <param name="instrumentName">The instrument name.</param>
         /// <returns>Boolean indicating if the instrument is valid.</returns>
-        internal static bool IsValidInstrumentName(string instrumentName)
+        public static bool IsValidInstrumentName(string instrumentName)
         {
             if (string.IsNullOrWhiteSpace(instrumentName))
             {
@@ -45,7 +62,7 @@ namespace OpenTelemetry.Metrics
         /// <remarks>See specification: <see href="https://github.com/open-telemetry/opentelemetry-specification/blob/main/specification/metrics/api.md#instrument"/>.</remarks>
         /// <param name="customViewName">The view name.</param>
         /// <returns>Boolean indicating if the instrument is valid.</returns>
-        internal static bool IsValidViewName(string customViewName)
+        public static bool IsValidViewName(string customViewName)
         {
             // Only validate the view name in case it's not null. In case it's null, the view name will be the instrument name as per the spec.
             if (customViewName == null)
@@ -55,7 +72,5 @@ namespace OpenTelemetry.Metrics
 
             return InstrumentNameRegex.IsMatch(customViewName);
         }
-
-        internal MeterProvider BuildSdk() => this.Build();
     }
 }